:recycle: Split app bar out fo default layout.
Why? code readability.
